Project, estimate, requirement ja task -entiteettien kenttien muuttaminen integeristä floattiin.

+++ src/main/java/com/tol/tenderwork/domain/Estimate.java
@@ -30,53 +30,53 @@ public class Estimate implements Serializable {
     @Max(value = 31)
     @Column(name = "workdays_in_month", nullable = false)
     private Integer workdaysInMonth;
-    
+
     @NotNull
     @Min(value = 1)
     @Column(name = "desired_project_duration", nullable = false)
     private Integer desiredProjectDuration;
-    
+
     @NotNull
     @Min(value = 1)
     @Column(name = "daily_price", nullable = false)
     private Long dailyPrice;
-    
+
     @NotNull
     @Min(value = 0)
     @Max(value = 1)
     @Column(name = "specification_factor", nullable = false)
     private Float specificationFactor;
-    
+
     @NotNull
     @Min(value = 0)
     @Max(value = 1)
     @Column(name = "testing_factor", nullable = false)
     private Float testingFactor;
-    
+
     @NotNull
     @Min(value = 0)
     @Max(value = 1)
     @Column(name = "implementation_factor", nullable = false)
     private Float implementationFactor;
-    
+
     @NotNull
     @Min(value = 0)
     @Max(value = 1)
     @Column(name = "synergy_benefit", nullable = false)
     private Float synergyBenefit;
-    
+
     @Column(name = "total_price")
     private Long totalPrice;
-    
+
     @Column(name = "total_duration")
-    private Integer totalDuration;
-    
+    private Float totalDuration;
+
     @Column(name = "resourcing")
-    private Integer resourcing;
-    
+    private Float resourcing;
+
     @Column(name = "total_synergy_benefit")
-    private Integer totalSynergyBenefit;
-    
+    private Float totalSynergyBenefit;
+
     @ManyToOne
     @JoinColumn(name = "created_by_id")
     private User createdBy;
@@ -101,7 +101,7 @@ public class Estimate implements Serializable {
     public Integer getWorkdaysInMonth() {
         return workdaysInMonth;
     }
-    
+
     public void setWorkdaysInMonth(Integer workdaysInMonth) {
         this.workdaysInMonth = workdaysInMonth;
     }
@@ -109,7 +109,7 @@ public class Estimate implements Serializable {
     public Integer getDesiredProjectDuration() {
         return desiredProjectDuration;
     }
-    
+
     public void setDesiredProjectDuration(Integer desiredProjectDuration) {
         this.desiredProjectDuration = desiredProjectDuration;
     }
@@ -117,7 +117,7 @@ public class Estimate implements Serializable {
     public Long getDailyPrice() {
         return dailyPrice;
     }
-    
+
     public void setDailyPrice(Long dailyPrice) {
         this.dailyPrice = dailyPrice;
     }
@@ -125,7 +125,7 @@ public class Estimate implements Serializable {
     public Float getSpecificationFactor() {
         return specificationFactor;
     }
-    
+
     public void setSpecificationFactor(Float specificationFactor) {
         this.specificationFactor = specificationFactor;
     }
@@ -133,7 +133,7 @@ public class Estimate implements Serializable {
     public Float getTestingFactor() {
         return testingFactor;
     }
-    
+
     public void setTestingFactor(Float testingFactor) {
         this.testingFactor = testingFactor;
     }
@@ -141,7 +141,7 @@ public class Estimate implements Serializable {
     public Float getImplementationFactor() {
         return implementationFactor;
     }
-    
+
     public void setImplementationFactor(Float implementationFactor) {
         this.implementationFactor = implementationFactor;
     }
@@ -149,7 +149,7 @@ public class Estimate implements Serializable {
     public Float getSynergyBenefit() {
         return synergyBenefit;
     }
-    
+
     public void setSynergyBenefit(Float synergyBenefit) {
         this.synergyBenefit = synergyBenefit;
     }
@@ -157,32 +157,32 @@ public class Estimate implements Serializable {
     public Long getTotalPrice() {
         return totalPrice;
     }
-    
+
     public void setTotalPrice(Long totalPrice) {
         this.totalPrice = totalPrice;
     }
 
-    public Integer getTotalDuration() {
+    public Float getTotalDuration() {
         return totalDuration;
     }
-    
-    public void setTotalDuration(Integer totalDuration) {
+
+    public void setTotalDuration(Float totalDuration) {
         this.totalDuration = totalDuration;
     }
 
-    public Integer getResourcing() {
+    public Float getResourcing() {
         return resourcing;
     }
-    
-    public void setResourcing(Integer resourcing) {
+
+    public void setResourcing(Float resourcing) {
         this.resourcing = resourcing;
     }
 
-    public Integer getTotalSynergyBenefit() {
+    public Float getTotalSynergyBenefit() {
         return totalSynergyBenefit;
     }
-    
-    public void setTotalSynergyBenefit(Integer totalSynergyBenefit) {
+
+    public void setTotalSynergyBenefit(Float totalSynergyBenefit) {
         this.totalSynergyBenefit = totalSynergyBenefit;
     }
